Slightly restructure the `datastore` API again, everything important should now be exposed and `datastore.core.*` only needed by internal stuff
