+++ datastore/filesystem/__init__.py
@@ -1,5 +1,6 @@
-__version__ = '1.2'
-__author__ = 'Juan Batiz-Benet <juan@benet.ai>'
+__version__ = "2.0"
+__author__ = "Juan Batiz-Benet, Alexander Schlarb"
+__email__ = "juan@benet.ai, alexander@ninetailed.ninja"
 __doc__ = """
 Filesystem datastore implementation
 """
