+++ datastore/typing.py
@@ -1,0 +1,2 @@
+from .core.util.stream import ArbitraryReceiveChannel
+from .core.util.stream import ArbitraryReceiveStream
