+++ datastore/util.py
@@ -1,0 +1,2 @@
+from .core.util.stream import receive_channel_from
+from .core.util.stream import receive_stream_from
